[DEV] Increase number of buttons for joystick to 64

+++ Middlewares/ST/STM32_USB_Device_Library/Class/HID/Inc/usbd_hid.h
@@ -119,7 +119,7 @@ typedef struct
 typedef struct
 {
     uint8_t report_id;
-    uint32_t buttons;
+    uint64_t buttons;
     int8_t x;
     int8_t y;
     int8_t z;
